Cleaned up dal compile-time warnings.
git-svn-id: http://usg.lofar.org/svn/code/trunk/src/DAL@1644 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ implement/TBB.h
@@ -119,7 +119,7 @@ namespace DAL {
     Int16 spare;
     UInt16 crc;
   };
-  long seqnrLast;
+  UInt32 seqnrLast;
   bool bigendian;
   time_t sample_time;  // For date
   string name;
